Dedicated namespace for parameter utils.
